% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{feather}
\alias{feather}
\alias{read_feather}
\alias{write_feather}
\title{Read and write feather files.}
\usage{
read_feather(path)

write_feather(x, path)
}
\arguments{
\item{path}{Path to feather file}
}
\value{
Both functions return a tibble/data frame. \code{read_feather}
  returns the input from disk. \code{write_feather} invisible returns
  \code{x} (so you can use this function in a pipeline).
}
\description{
Read and write feather files.
}

