% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather.R
\name{feather_metadata}
\alias{feather_metadata}
\title{Retrieve metadata about a feather file}
\usage{
feather_metadata(path)
}
\arguments{
\item{path}{Path to feather file}
}
\value{
A list with class "feather_metadata".
}
\description{
Returns the dimensions, field names, and types; and optional dataset
description.
}
