% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{feather}
\alias{$.feather}
\alias{[.feather}
\alias{[[.feather}
\alias{as.data.frame.feather}
\alias{as_data_frame.feather}
\alias{dim.feather}
\alias{dimnames.feather}
\alias{feather}
\alias{print.feather}
\alias{row.names.feather}
\title{Access a feather store like a data frame}
\usage{
feather(path)

\method{row.names}{feather}(x)

\method{dimnames}{feather}(x)

\method{dim}{feather}(x)

\method{[[}{feather}(x, i, exact = TRUE)

\method{$}{feather}(x, i)

\method{[}{feather}(x, i, j, drop = FALSE)

\method{as.data.frame}{feather}(x, row.names = NULL, optional = FALSE, ...)

\method{as_data_frame}{feather}(x, ...)

\method{print}{feather}(x, ...)
}
\arguments{
\item{path}{Path to feather file}

\item{x}{A data frame to write to disk}
}
\value{
An object of class \code{feather}
}
\description{
These functions permit using a feather dataset much like a regular
(read-only) data frame.
}

